# api/api.yaml
...
openapi: 3.0.3
info:
  title: codziennemotto.pl REST API
  description: codziennemotto.pl REST API
  version: 1.0.0
servers:
  - url: 'https'
paths:
...
  /text-set/{setId}/join-link/{joinLinkId}:
    delete:
      description: Delete JoinLink.
      parameters:
        - in: path
          schema:
            type: integer
            description: TextSet's ID
          name: setId
        - in: path
          schema:
            type: integer
          name: joinLinkId
          description: JoinLink's ID
      tags:
        - JoinLink
      responses:
        '204':
          description: Deleted JoinLink
        '403':
          description: Forbidden - no valid authentication.
        '400':
          description: Bad Request - You're not owner of TextSet's, or JoinLink is not exist.
  /text-set/create-new:
    description: Create new TextSet
    post:
      operationId: pl.codziennemotto.controllers.textset.TextSetController#createNewTextSetEndpoint
      tags:
        - TextSet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTextSetPayload'

      responses:
        '200':
          description: TextSet successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextSet'
        '403':
          description: Forbidden - no valid authentication
  /text-set/{id}:
    description: Get TextSet by ID
    delete:
      tags:
        - TextSet
      parameters:
        - in: path
          schema:
            type: integer
          name: id

      responses:
        '204':
          description: Deleted
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: bad request - you're not owner of target TextSet, or TextSet doesn't exist.
    get:
      tags:
        - TextSet
      parameters:
        - in: path
          schema:
            type: integer
          name: id

      description: You must be authenticated as TextSet's owner or Reader.
      operationId: pl.codziennemotto.controllers.textset.TextSetController#textSetByIdEndpoint
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TextSet"
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: bad request - you're not owner or reader of target TextSet, or TextSet doesn't exist.
  /text-set/{id}/create-join-link:
    description: Creates Join Link
    post:
      tags:
      - JoinLink
      parameters:
        - in: path
          schema:
            type: integer
          name: id
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinLink'
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: bad request - You're not owner of the TextSet or it doesn't exist.
  /text-set/{id}/readers/{readerId}:
    description: Deletes Reader
    delete:
      tags:
        - TextSet
      parameters:
        - in: path
          schema:
            type: integer
          name: id
        - in: path
          schema:
            type: integer
          name: readerId
      responses:
        '204':
          description: 'Deleted'
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: bad request - you're not owner of target TextSet, or TextSet [| or Reader] doesn't exist.
  /text-set/{id}/join-links:
    description: Gets all JoinLink's in TextSet [even expired]
    post:
      tags:
        - JoinLink
      parameters:
        - in: path
          schema:
            type: integer
          name: id
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JoinLink'

        '403':
          description: Forbidden - no valid authentication
        '400':
          description: bad request - You're not owner of the TextSet or it doesn't exist.
  /text-set/{id}/quit:
    description: Reader leaves the TextSet.
    delete:
      tags:
        - TextSet
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: TextSet's id
      responses:
        '204':
          description: 'Successfully deleted.'
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: Bad request - you're not a Reader of the TextSet
  /text-set/{id}/readers:
    description: Get TextSet's Readers
    get:
      tags:
        - TextSet
      parameters:
        - in: path
          schema:
            type: integer
          name: id
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reader"
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: bad request - You're not owner of the TextSet or it doesn't exist.
  /text-set/{id}/texts/all:
    description: Get all TextSet`s text. Only for TextSet's owner
    get:
      tags:
        - TextSet
      parameters:
        - in: path
          schema:
            type: integer
          name: id
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Text"
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: bad request - You're not owner of the TextSet or it doesn't exist.
  /text-set/{id}/texts/all/visible:
    description: Get all TextSet`s text where authorized user is allowed to see.
    get:
      tags:
        - TextSet
      parameters:
        - in: path
          schema:
            type: integer
          name: id
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Text"
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: bad request - You're not allowed to access the TextSet.
  /text-set/{id}/texts/past:
    description: Get past TextSet`s text.
    get:
      tags:
        - TextSet
      parameters:
        - in: path
          schema:
            type: integer
          name: id
      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Text"
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: bad request - You're not reader or owner of the TextSet or it doesn't exist.
  /text-set/{id}/add:
    put:
      responses:
        '200':
          description: Successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Text"
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: bad request - You're not owner of the TextSet or it doesn't exist
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTextPayload"
      tags:
        - TextSet
      parameters:
        - in: path
          schema:
            type: integer
          name: id
  /text-set/join-with-code/{code}:
    post:
      tags:
        - JoinLink
      parameters:
        - in: path
          name: code
      responses:
        '200':
          description: OK - Reader created.
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/Reader"
        '403':
          description: Forbidden - No valid authentication.
        '400':
          description: Bad request - probably there's no code in database.
...
  /text-set/text/by-id/{textId}:
    get:
      parameters:
        - in: path
          schema:
            type: integer
          name: textId
          description: Text's ID
      tags:
        - TextSet
      responses:
        '200':
          description: Successfully fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Text"
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: You're not allowed to see this Text.
  /text-set/{setId}/{textId}:
    get:
      parameters:
        - in: path
          schema:
            type: integer
          name: setId
        - in: path
          schema:
            type: integer
          name: textId
      tags:
        - TextSet
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Text"
          description: Success
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: Something wrong, target Text must be accessible for you [If you're reader, Text must be past]
    delete:
      parameters:
        - in: path
          schema:
            type: integer
          name: setId
        - in: path
          schema:
            type: integer
          name: textId
      tags:
        - TextSet
      responses:
        '204':
          description: Successfully deleted.
        '403':
          description: Forbidden - no valid authentication
        '400':
          description: You're not owner of the TextSet or it doesn't exist.
  /text-set/where-i-am-owner:
    get:
      description: 'Returns list of TextSet where authenticated user is owner'
      tags:
        - TextSet
      responses: 
        '200':
          description: Successfully obtained
          content: 
            application/json:
              schema:
                type: array
                items: 
                  $ref: "#/components/schemas/TextSet"
        '403':
          description: Forbidden - no valid authentication
  /text-set/where-i-am-reader:
    get:
      description: 'Returns list of TextSet where authenticated user is reader'
      tags:
        - TextSet
      responses: 
        '200':
          description: Successfully obtained
          content: 
            application/json:
              schema:
                type: array
                items: 
                  $ref: "#/components/schemas/TextSet"
        '403':
          description: Forbidden - no valid authentication
  /summary: 
    get:
      description: Returns summary of all list's that authenticated user has access
      tags:
        - TextSet
      responses:
        '200': 
          description: Successfully obtained
          content: 
            application/json:
              schema: 
                type: array
                items: 
                  $ref: "#/components/schemas/SummaryObject"
        '403': 
          description: Forbidden - no valid authentication.
...
  /text-set/{id}/join-with-code/{code}:
    post:
      tags: 
        - JoinLink
        
      parameters:
        - in: path
          schema:
            type: string
          name: code
        - in: path
          schema:
            type: integer
          name: id
      
      responses: 
        '200':
          description: Successfully joined to TextSet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reader"
        '403': 
          description: Forbidden - no valid authentication.
        '400': 
          description: Bad request - no matching  JoinLink in database
  /text-set/{id}/readers/include-users:
    get:
      tags:
        - TextSet

      parameters:
        - in: path
          name: id
          schema:
            type: string
      responses:
        '200':
          description: Successfully fetch
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReaderIncludeUser"
        '403':
          description: Forbidden - no valid authentication.
        '400':
          description: bad request - no matching data corresponding to your authentication.
components:
  schemas:
...
    TextSet:
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        ownerId:
          type: integer
          nullable: true
    Reader:
      properties:
        id:
          type: integer
        text_set_id:
          type: integer
          nullable: true
        user_id:
          type: integer
    Text:
      properties:
        id:
          type: integer
        order:
          type: integer
        shown:
          type: string
          format: "yyyy-MM-dd"
          nullable: true
        text_set_id:
          type: integer
          nullable: true
        text:
          type: string
    JoinLink:
      type: object
      title: "JoinLink"
      properties:
        id:
          type: integer
        text_set_id:
          type: integer
          nullable: true
        activeUntil:
          type: string
          format: "yyyy-MM-dd"
        code:
          type: string
...
    CreateTextSetPayload:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
    AddTextPayload:
      properties:
        text:
          type: string
        date:
          type: string
          format: "yyyy-MM-dd"
        order:
          type: integer
...
    SummaryObject:
      type: object
      properties:
        textSet:
          $ref: '#/components/schemas/TextSet'
        text:
          type: string
      required:
        - textSet
    ReaderIncludeUser:
      type: object
      properties:
        userName:
          type: string
        userId:
          type: integer
        reader:
          $ref: "#/components/schemas/Reader"